# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: WAVE / AutoWash Pro API
  version: 0.0.1
  description: REST API for the WAVE car-wash platform (Express backend).
servers:
  - url: /api
    description: Global API prefix
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          example: Validation failed
        error:
          type: string
          example: Bad Request
...
    PaginationMeta:
      type: object
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        total: { type: integer, example: 42 }
        totalPages: { type: integer, example: 3 }
...
    CreateStaffShift:
      description: >-
        Anonymous capacity-based shift: date + block + capacity. No staff,
        no station. `fullday` creates two shift records (morning + afternoon).
      type: object
      required: [date, block]
      properties:
        date: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$', example: '2026-06-01' }
        block: { type: string, description: 'morning | afternoon | fullday', example: 'morning' }
        capacity: { type: integer, minimum: 1, maximum: 10, default: 1, description: 'Concurrent washes bookable in this shift.', example: 2 }
        note: { type: string, maxLength: 500, example: 'Morning shift' }
...
    StaffShiftResponse:
      type: object
      properties:
        id: { type: string, example: '6601e3b3f1a2c3a4b5d6e7f8' }
        staffId: { type: string, description: 'Legacy per-staff shifts only.', example: '6601e3b3f1a2c3a4b5d6e7f8' }
        shiftType: { type: string, example: 'washer' }
        stationName: { type: string, description: 'Legacy field — no longer settable.', example: 'Bay 1' }
        capacity: { type: integer, description: 'Concurrent washes bookable in this shift.', example: 2 }
        startAt: { type: string, format: date-time, example: '2026-06-01T08:00:00.000Z' }
        endAt: { type: string, format: date-time, example: '2026-06-01T12:00:00.000Z' }
        status: { type: string, description: 'Time-derived (scheduled/active/completed); only cancelled is stored manually.', example: 'scheduled' }
        note: { type: string, example: 'Morning shift' }
    StaffShiftListResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/StaffShiftResponse' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }
...
paths:
...
  /admin/shifts:
    get:
      tags: [shifts]
      summary: List staff shifts (admin/manager, paginated)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: staffId, in: query, schema: { type: string } }
        - { name: shiftType, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
        - { name: startFrom, in: query, schema: { type: string, format: date-time } }
        - { name: startTo, in: query, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaffShiftListResponse' }
    post:
      tags: [shifts]
      summary: Create an anonymous capacity shift (admin/manager). fullday returns two shifts. Slots become bookable immediately.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateStaffShift' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: '#/components/schemas/StaffShiftResponse' }
                  - type: array
                    items: { $ref: '#/components/schemas/StaffShiftResponse' }
        '400':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...